fix error enter ID

+++ app/Http/Controllers/RUEmployerController.php
@@ -13,8 +13,11 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\File;
 use Carbon\Carbon;
+use App\Models\Job_posting;
 use App\Http\Requests\Employer\UpdateRequest;
 use App\Http\Requests\Employer\ChangePasswordRequest;
+
+use function Psy\debug;
 
 class RUEmployerController extends Controller
 {
@@ -57,8 +60,13 @@ class RUEmployerController extends Controller
      */
     public function show($id)
     {
-        $showEmploy = Employer::find($id);
-        return view('DashboardTemplate.employer.detail_employer', compact('showEmploy'));
+        $id = Auth::user()->employer->id;
+        $showEmploy = DB::table('employers')->where('id', $id)->first();
+        if ($showEmploy == null) {
+            return view('error.404');
+        } else {
+            return view('DashboardTemplate.employer.detail_employer', compact('showEmploy'));
+        }
     }
 
     /**
@@ -69,8 +77,13 @@ class RUEmployerController extends Controller
      */
     public function edit($id)
     {
-        $showEmploy = Employer::find($id);
-        return view('DashboardTemplate.employer.edit_employer', compact('showEmploy'));
+        $id = Auth::user()->employer->id;
+        $showEmploy = DB::table('employers')->where('id', $id)->first();
+        if ($showEmploy == null) {
+            return view('error.404');
+        } else {
+            return view('DashboardTemplate.employer.edit_employer', compact('showEmploy'));
+        }
     }
 
     /**
@@ -82,19 +95,33 @@ class RUEmployerController extends Controller
      */
     public function update(UpdateRequest $request, $id)
     {
-        $Employer = Employer::find($id);
-        if ($request->hasFile('image_upload')) {
-            $file = $request->file('image_upload');
-            $file_name = $file->getClientOriginalName();
-            $file->move('img', $file_name);
-            $imagePath = public_path('img/' . $Employer->image);
-            if (File::exists($imagePath)) {
-                unlink($imagePath);
+        $check = DB::table('employers')->where('id',$id)->first();
+        if($check == null){
+            return view('error.404');
+        }else{
+            $Employer = Employer::find($id);
+            if ($request->hasFile('image_upload')) {
+                $file = $request->file('image_upload');
+                $file_name = $file->getClientOriginalName();
+                $file->move('img', $file_name);
+                $imagePath = public_path('img/' . $Employer->image);
+                if (File::exists($imagePath)) {
+                    unlink($imagePath);
+                }
+                    $Employer->image = $file_name;
             }
-            $Employer->image = $file_name;
+            $Employer->update([
+                'website' => htmlspecialchars($request->website),
+                'infor' =>htmlspecialchars($request->infor),
+                'responsibility' =>htmlspecialchars($request->responsibility),
+                'welfare' =>htmlspecialchars($request->welfare),
+                'name_company' =>htmlspecialchars($request->name_company),
+                'address' => htmlspecialchars($request->address),
+                'email' =>htmlspecialchars($request->email),
+                'phone_number' =>htmlspecialchars($request->phone_number),
+            ]);
+            return redirect()->back()->with('notify', 'Update Susscessfully');
         }
-        $Employer->update($request->all());
-        return redirect()->back()->with('notify', 'Update Susscessfully');
     }
 
     /**
@@ -105,49 +132,81 @@ class RUEmployerController extends Controller
      */
     public function destroy($id)
     {
-        $re = DB::table('recruitments')->select('id')->where('customer_id', '=', $id)->get();
-        Recruitment::destroy($re[0]->id);
-        return redirect()->route('CRUDJobByEmployer.show',Auth::user()->employer->id)->with('notify', 'Delete Susscessfully');
+        $check = DB::table('recruitments')->where('id', $id)->first();
+        if ($check == null) {
+            return view('error.404');
+        } else {
+            $re = DB::table('recruitments')->select('id')->where('customer_id', '=', $id)->get();
+            Recruitment::destroy($re[0]->id);
+            return redirect()->route('CRUDJobByEmployer.show', Auth::user()->employer->id)->with('notify', 'Delete Susscessfully');
+        }
     }
 
     public function detail_recruitment($id)
     {
-        $customer = Customer::find($id);
-        $re = DB::table('recruitments')->select('id')->where('customer_id', '=', $id)->get();
-        $update_status = Recruitment::find($re[0]->id);
-        $update_status->update([
-            'status' => '1',
-        ]);
-        return view('DashboardTemplate.Job_postings.detail', compact('customer'));
+        $check = DB::table('job_postings')->where('employer_id', Auth::user()->employer->id)->first();
+        $check = Job_posting::findOrFail($check->id);
+        $re = $check->customers;
+        $array = array();
+        for ($i = 0; $i < count($re); $i++) {
+            $array[] = $re[$i]->id;
+        }
+        if (in_array($id, $array)) {
+            $customer = DB::table('customers')->where('id', $id)->first();
+            $re = DB::table('recruitments')->select('id')->where('customer_id', '=', $id)->get();
+            $update_status = Recruitment::find($re[0]->id);
+            $update_status->update([
+                'status' => '1',
+            ]);
+            return view('DashboardTemplate.Job_postings.detail', compact('customer'));
+        } else {
+            return view('error.404');
+        }
     }
+
     public function recruit(Customer $customer)
     {
-        $id = Customer::find($customer->id);
-        $job = $id->job_postings->take(1);
-        $tomorrow = Carbon::tomorrow();
-        Mail::send('DashboardTemplate.emails.send', compact('customer', 'job','tomorrow'), function ($email) use ($customer) {
-            $email->subject('Thông báo tuyển dụng');
-            $email->to($customer->email);
-        });
-        return redirect()->back()->with('notify', 'Send mail susscrssfully !');
+        $id = DB::table('recruitments')->where('customer_id',$customer->id)->first();
+        if($id == null){
+            return view('error.404');
+        }else{
+            $id = Customer::findOrFail($customer->id);
+            $job = $id->job_postings->take(1);
+            $tomorrow = Carbon::tomorrow();
+            Mail::send('DashboardTemplate.emails.send', compact('customer', 'job', 'tomorrow'), function ($email) use ($customer) {
+                $email->subject('Thông báo tuyển dụng');
+                $email->to($customer->email);
+            });
+            return redirect()->back()->with('notify', 'Send mail susscrssfully !');
+        }
     }
 
     public function changepass(ChangePasswordRequest $request, $id)
     {
-        if (!Hash::check($request->old_pass, auth()->user()->password)) {
-            return redirect()->back()->with("error", "Đổi mật khẩu thất bại!");
+        $id = DB::table('users')->where('id', $id)->first();
+        if ($id == null) {
+            return view('error.404');
         } else {
-            $change = User::find($id);
-            $change->update([
-                'password' => $request->new_pass,
-            ]);
-            Auth::logout();
-            return redirect()->route('login')->with("status", "Đổi mật khẩu thành công, Vui lòng đăng nhập lại!");
+            if (!Hash::check($request->old_pass, auth()->user()->password)) {
+                return redirect()->back()->with("error", "Đổi mật khẩu thất bại!");
+            } else {
+                $change = User::find($id);
+                $change->update([
+                    'password' => $request->new_pass,
+                ]);
+                Auth::logout();
+                return redirect()->route('login')->with("status", "Đổi mật khẩu thành công, Vui lòng đăng nhập lại!");
+            }
         }
     }
     public function showlayout($id)
     {
-        $showEmployByID = User::find($id);
-        return view('DashboardTemplate.employer.showlayout', compact('showEmployByID'));
+        $id = Auth::user()->id;
+        $showEmployByID = DB::table('users')->where('id', $id)->first();
+        if ($showEmployByID == null) {
+            return view('error.404');
+        } else {
+            return view('DashboardTemplate.employer.showlayout', compact('showEmployByID'));
+        }
     }
 }
